refactor(simplegeom): :sparkles: 更新文件顺序，新增transform, length方法
1. transform用于地理坐标系和笛卡尔坐标系（北东地）转换
2. length用于求长度或周长

BREAKING CHANGE: closest_point改名为cloest_points函数
